bundle all helper scripts into one
