+++ workflow/snakefile/prep_ksnp_snakefile
@@ -9,11 +9,12 @@ current_directory = os.getcwd()
 # Load the JSON data
 with open('{}/config.json'.format(current_directory)) as f:
     data = json.load(f)
-
+clean_fasta_dir = data["clean_data_dir"]
+work_data_dir = data["work_data_dir"]
 rule_all_list = [
-                "{}/clean_fastas_complete.txt".format(current_directory),
-                "{}/clean_fastas/ksnp4_input_file.txt".format(current_directory),
-                "{}/clean_fastas/Kchooser4_ksnp4_input_file.report".format(current_directory)
+                "{}/clean_fastas_complete.txt".format(work_data_dir),
+                "{}/ksnp4_input_file.txt".format(clean_fasta_dir),
+                "{}/Kchooser4_ksnp4_input_file.report".format(clean_fasta_dir)
                 ]
 
 rule all:
@@ -23,19 +24,12 @@ rule all:
 
 rule remove_special_characters_from_fasta_names:
     input:
-        raw_fasta_dir = data["raw_fasta_dir"],
         config = '{}/config.json'.format(current_directory)
-    params:
-        clean_fasta_dir = "{}/clean_fastas".format(current_directory)
     output:
-        touchpoint = "{}/clean_fastas_complete.txt".format(current_directory)
+        touchpoint = "{}/clean_fastas_complete.txt".format(work_data_dir)
     shell:
             """
-            mkdir -p {params.clean_fasta_dir}
-
-            clean_fasta_filenames \
-                {input.raw_fasta_dir} \
-                {params.clean_fasta_dir} \
+            whole_genome_snp_utils clean-fasta-filenames \
                 {input.config}
 
             touch {output.touchpoint}
@@ -43,33 +37,34 @@ rule remove_special_characters_from_fasta_names:
 
 rule write_kSNP4_input_file:
     input:
-        touchpoint = "{}/clean_fastas_complete.txt".format(current_directory)
+        touchpoint = "{}/clean_fastas_complete.txt".format(work_data_dir)
     params:
-        clean_fasta_dir = directory("{}/clean_fastas".format(current_directory))
+        # clean_fasta_dir = directory("{}/clean_fastas".format(clean_fasta_dir))
+        clean_fasta_dir_ = clean_fasta_dir
     output:
-        ksnp_in_file = "{}/clean_fastas/ksnp4_input_file.txt".format(current_directory) # writing to fasta dir because Kchooser4 is picky 
+        ksnp_in_file = "{}/ksnp4_input_file.txt".format(clean_fasta_dir) # writing to fasta dir because Kchooser4 is picky 
     shell:
             """
-            MakeKSNP4infile -indir {params.clean_fasta_dir} -outfile {output.ksnp_in_file}
+            MakeKSNP4infile -indir {params.clean_fasta_dir_} -outfile {output.ksnp_in_file}
             """
 
 rule run_kchooser:
     input:
-        ksnp_in_file = "{}/clean_fastas/ksnp4_input_file.txt".format(current_directory)
+        ksnp4_in_file = "{}/ksnp4_input_file.txt".format(clean_fasta_dir)
     params:
-        clean_fasta_dir = directory("{}/clean_fastas".format(current_directory)),
+        clean_fasta_dir_ = clean_fasta_dir,
         current_directory_ = directory(current_directory),
-        ksnp4_input_file = "ksnp4_input_file.txt"
+        ksnp4_in_file = "ksnp4_input_file.txt" # using this relative path because kSNP4 is picky
     output:
-        output_from_kchooser = "{}/clean_fastas/Kchooser4_ksnp4_input_file.report".format(current_directory),
-        optimum_k_txt = "{}/optimum_k.txt".format(current_directory)
+        output_from_kchooser = "{}/Kchooser4_ksnp4_input_file.report".format(clean_fasta_dir),
+        optimum_k_txt = "{}/optimum_k.txt".format(work_data_dir)
     shell:
         """
-        cd {params.clean_fasta_dir}
+        cd {params.clean_fasta_dir_}
 
-        Kchooser4 -in {params.ksnp4_input_file}
+        Kchooser4 -in {params.ksnp4_in_file}
 
         cd {params.current_directory_}
 
-        parse_kchooser_report {output.output_from_kchooser} > {output.optimum_k_txt}
+        whole_genome_snp_utils parse-kchooser-report {output.output_from_kchooser} > {output.optimum_k_txt}
         """
